Move python-lambda-layer
ESLint ignore for beforeBinding command
Run prettier --write for bssh tools lambda layer
