+++ lib/workload/stateless/statelessStackCollectionClass.ts
@@ -14,8 +14,10 @@ import {
   SequenceRunManagerStack,
   SequenceRunManagerStackProps,
 } from './stacks/sequence-run-manager/deploy/stack';
-import { BsRunsUploadManagerStackProps } from './stacks/bs-runs-upload-manager/deploy/stack';
-import { BsRunsUploadManagerStack } from './stacks/bs-runs-upload-manager/deploy/stack';
+import {
+  BsRunsUploadManagerStack,
+  BsRunsUploadManagerStackProps,
+} from './stacks/bs-runs-upload-manager/deploy/stack';
 
 export interface StatelessStackCollectionProps {
   postgresManagerStackProps: PostgresManagerStackProps;
